usb 0825 commit

+++ model/usb/host/include/usb_pnp_notify.h
@@ -62,7 +62,7 @@ struct UsbInfoQueryPara {
     UsbInfoQueryParaType type;
     union {
         int32_t id;
-        uint32_t usbDevAddr;
+        uintptr_t usbDevAddr;
         struct {
             int32_t devNum;
             int32_t busNum;
